вымвыммм

+++ src/main/java/searchengine/model/Site.java
@@ -12,7 +12,7 @@ public class Site {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id; // Note that this is Long
+    private Integer id; // Изменено на Integer
 
     @Enumerated(EnumType.STRING)
     @Column(columnDefinition = "ENUM('INDEXING', 'INDEXED', 'FAILED')", nullable = false)
