+++ src/main/java/searchengine/repositories/PageRepository.java
@@ -4,13 +4,10 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import searchengine.model.Page;
 import searchengine.model.Site;
 
-public interface PageRepository extends JpaRepository<Page, Long> { // Keep Long for ID
-    // Удаление страниц по идентификатору сайта
-    void deleteBySiteId(Long siteId); // Use Long
+public interface PageRepository extends JpaRepository<Page, Integer> { // Изменено на Integer
+    void deleteBySiteId(Integer siteId); // Изменено на Integer
 
-    // Проверка, существуют ли страницы для данного сайта
-    boolean existsBySiteId(Long siteId); // Use Long
+    boolean existsBySiteId(Integer siteId); // Изменено на Integer
 
-    // Проверка, существует ли страница для данного сайта и пути
     boolean existsBySiteAndPath(Site site, String path);
 }
