+++ src/main/java/searchengine/repositories/SiteRepository.java
@@ -3,9 +3,6 @@ package searchengine.repositories;
 import org.springframework.data.jpa.repository.JpaRepository;
 import searchengine.model.Site;
 
-public interface SiteRepository extends JpaRepository<Site, Long> {
-    // Find a Site by its URL
+public interface SiteRepository extends JpaRepository<Site, Integer> { // Изменено на Integer
     Site findByUrl(String url);
-
-    // Additional custom methods can be added as needed
 }
